# docs/swagger.yaml
basePath: /api
definitions:
...
  agent.ErrorResponse:
    properties:
      error:
        example: workload not found
        type: string
    type: object
...
  agent.TunnelRequest:
    properties:
      token:
        example: eyJhIjoiNjA2...
        type: string
    type: object
  agent.TunnelStatus:
    properties:
      configured:
        example: true
        type: boolean
      running:
        example: true
        type: boolean
    type: object
...
info:
  contact:
    name: Jetty Support
    url: https://github.com/ncwardell/jetty
  description: P2P Docker Compose orchestration with Cloudflare WARP mesh networking
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Jetty API
  version: 0.0.x
paths:
...
  /tunnel:
    delete:
      description: Default scope=node stops this node's cloudflared connector and
        leaves the cluster token and every other node untouched - use this to pull
        one node out of Cloudflare's load-balancing rotation. scope=cluster clears
        the token everywhere and stops every connector, taking public traffic down
        until a token is set again.
      parameters:
      - description: node (default) or cluster
        in: query
        name: scope
        type: string
      - description: Target node id or name
        in: query
        name: node
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.TunnelStatus'
        "400":
          description: Invalid scope
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
      summary: Detach a node from the tunnel, or remove it cluster-wide
      tags:
      - tunnel
    get:
      description: Returns Cloudflare tunnel status. "configured" is cluster-wide
        (a token is set); "enabled" and "running" are local to this node. Pass ?node=<id|name>
        to query a peer.
      parameters:
      - description: Target node id or name
        in: query
        name: node
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.TunnelStatus'
      summary: Get tunnel status
      tags:
      - tunnel
    post:
      consumes:
      - application/json
      description: With scope=cluster (or a body containing a token), sets the Cloudflare
        tunnel token and propagates it to all nodes. With the default scope=node and
        no token, re-attaches this node to the existing cluster token after a scope=node
        delete.
      parameters:
      - description: node (default) or cluster
        in: query
        name: scope
        type: string
      - description: Target node id or name
        in: query
        name: node
        type: string
      - description: Tunnel token (required for scope=cluster)
        in: body
        name: token
        schema:
          $ref: '#/definitions/agent.TunnelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.TunnelStatus'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/agent.ErrorResponse'
      summary: Configure tunnel or re-enable this node
      tags:
      - tunnel
...
schemes:
- https
- http
security:
- ApiKeyAuth: []
securityDefinitions:
  ApiKeyAuth:
    description: API key for authentication (JETTY_SECRET)
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"